backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Validation & Serialization
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0
factory-boy>=3.3.0

# Code Quality
ruff>=0.1.14
mypy>=1.8.0

# Utilities
python-dotenv>=1.0.0
